Move publication date interval filters to own component
